Update driver.flow.compose.json
German language added
I don't know if the lines 79, 90 and 101 work. I may have to correct these lines later.
I know 2 German users of the app (I myself do not have a LinkTap valve) who I can then ask.

Line 43: "When the valve in not broken" or "When the valve is not broken"?